[Refactor]: 검증 메서드 분리
